# src/auth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/departments": {
            "get": {
                "tags": [
                    "部门"
                ],
                "summary": "列出所有部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页尺寸",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"departments\":[{\"id\":2,\"name\":\"root\",\"parent_id\":null,\"created_at\":\"2021-01-21 15:10:26\",\"updated_at\":\"2021-01-21 15:10:26\"},{\"id\":5,\"name\":\"sub2\",\"parent_id\":2,\"created_at\":\"2021-01-21 15:11:07\",\"updated_at\":\"2021-01-21 15:11:07\"}],\"message\":\"Success\",\"page\":1,\"page_size\":50,\"pages\":1,\"total\":2}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "部门"
                ],
                "summary": "更新部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"department\":{\"id\":4,\"name\":\"sub_dept3\",\"parent_id\":2,\"created_at\":\"2021-01-21 15:11:00\",\"updated_at\":\"2021-01-21 15:11:00\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "部门"
                ],
                "summary": "新建部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"department\":{\"id\":4,\"name\":\"sub_dept3\",\"parent_id\":2,\"created_at\":\"2021-01-21 15:11:00\",\"updated_at\":\"2021-01-21 15:11:00\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/departments/tree": {
            "get": {
                "tags": [
                    "部门"
                ],
                "summary": "以树结构列出所有部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"tree\":{\"id\":2,\"name\":\"root\",\"sub_department\":[{\"id\":4,\"name\":\"sub1\",\"sub_department\":[],\"created_at\":\"2021-01-21 15:11:00\",\"updated_at\":\"2021-01-21 15:11:00\"},{\"id\":5,\"name\":\"sub2\",\"sub_department\":[],\"created_at\":\"2021-01-21 15:11:07\",\"updated_at\":\"2021-01-21 15:11:07\"},{\"id\":6,\"name\":\"sub3\",\"sub_department\":[{\"id\":9,\"name\":\"sub3_1\",\"sub_department\":[],\"created_at\":\"2021-02-19 03:17:30\",\"updated_at\":\"2021-02-19 03:17:33\"}],\"created_at\":\"2021-01-21 15:11:10\",\"updated_at\":\"2021-01-21 15:11:10\"},{\"id\":8,\"name\":\"sub1_1\",\"sub_department\":[],\"created_at\":\"2021-01-21 15:11:34\",\"updated_at\":\"2021-01-22 10:53:13\"}],\"created_at\":\"2021-01-21 15:10:26\",\"updated_at\":\"2021-01-21 15:10:26\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/departments/{department_id}": {
            "delete": {
                "tags": [
                    "部门"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/departments/{department_id}/tree": {
            "get": {
                "tags": [
                    "部门"
                ],
                "summary": "列出指定部门子树",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"tree\":{\"id\":6,\"name\":\"sub3\",\"sub_department\":[{\"id\":9,\"name\":\"sub3_1\",\"sub_department\":[],\"created_at\":\"2021-02-19 03:17:30\",\"updated_at\":\"2021-02-19 03:17:33\"}],\"created_at\":\"2021-01-21 15:11:10\",\"updated_at\":\"2021-01-21 15:11:10\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/departments/{department_id}/users": {
            "get": {
                "tags": [
                    "部门"
                ],
                "summary": "列出部门中所有用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"users\":[{\"id\":4,\"username\":\"test2\",\"is_owner\":false,\"joined_at\":\"2021-01-20 11:01:16\"},{\"id\":3,\"username\":\"test\",\"is_owner\":true,\"joined_at\":\"2021-01-20 11:01:32\"}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "部门"
                ],
                "summary": "添加用户至部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.UserDepartment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/departments/{department_id}/users/{user_id}": {
            "put": {
                "tags": [
                    "部门"
                ],
                "summary": "设置用户是否是部门Owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.IsOwnerForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "部门"
                ],
                "summary": "移除部门中用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.Department": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
...
        "database.UserDepartment": {
            "type": "object",
            "required": [
                "is_owner",
                "user_id"
            ],
            "properties": {
                "is_owner": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "form.IsOwnerForm": {
            "type": "object",
            "required": [
                "is_owner"
            ],
            "properties": {
                "is_owner": {
                    "type": "boolean"
                }
            }
        },
...
    }
}